v3.1.1
### Patch
 - Removed console warning for when slashCommandIdsToDelete is not used.

+++ src/handlers/loadCommands.js
@@ -118,8 +118,10 @@ module.exports = ({
   };
   (async () => {
     if (
-      slashCommandIdsToDelete.global.length > 0 ||
-      slashCommandIdsToDelete.guild.length > 0
+      (slashCommandIdsToDelete.global &&
+        slashCommandIdsToDelete.global.length > 0) ||
+      (slashCommandIdsToDelete.guild &&
+        slashCommandIdsToDelete.guild.length > 0)
     ) {
       await deleteSlashCommands();
     }
